requirements.txt
selenium>=4.15.0
webdriver-manager>=4.0.1
python-dotenv>=1.0.0
PyPDF2>=3.0.0
reportlab>=4.0.0
